temp commit

+++ src/main/java/it/unipi/mircv/File/DocumentIndexHandler.java
@@ -62,12 +62,12 @@ public class DocumentIndexHandler{
         return buffer.getInt();
     }
 
-    public int readAvgDocLen() throws IOException {
+    public float readAvgDocLen() throws IOException {
         ByteBuffer buffer = ByteBuffer.allocate(AVGDOCLENGHT_BYTES_LENGTH);
         fileChannel.position(0);
         fileChannel.read(buffer);
         buffer.position(0);
-        return buffer.getInt();
+        return buffer.getFloat();
     }
 
     public int collectionSize() throws IOException {
